Add tool to check the state of any BitTorrent peer.
btcon.py is a basic implementation of some parts of the BitTorrent protocol. This allows connecting to a peer and downloading the metadata. Further, the implementation can find out the download progress status of the peer.
bt-stat is a command line tool using btcon to connect to a peer given address and infohash, then prints out the torrent details and the peer progress status.